pyproject.toml
[project]
name = "stepml"
version = "0.1.0"
description = "Machine learning-based step chart analysis and difficulty rating system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
]

[dependency-groups]
# Development tools
dev = [
    "build",
    "pytest>=8.0.0",
    "ruff",
    "validate-pyproject[all]",
]

[project.scripts]
generate-baseline = "stepml.generate_baseline:generate_baseline"
generate-dataset = "stepml.generate_dataset:main"
generate-ratings = "stepml.generate_calculated_ratings:main"
generate-playlists = "stepml.generate_playlists:main"
train-models = "stepml.train_baseline_models:main"
sync-favorites = "stepml.sync_favorites:main"
analyze-performance = "stepml.analyze_performance_data:main"
